+++ src/shim/index.ts
@@ -3,6 +3,7 @@ import {ErrorCodes} from "../Constants";
 import SDKError from "../lib/SDKError";
 import * as Init from "./Initialize";
 import {checkSDKInitialized} from "./ShimUtils";
+export {deleteDeviceByPublicSigningKey} from "./Initialize";
 
 /**
  * Checks bowser functionality to ensure random number generation is supported.