+++ src/shim/sdk/UserSDK.ts
@@ -18,6 +18,7 @@ export function changePasscode(currentPasscode: string, newPasscode: string) {
         .map(() => undefined)
         .toPromise();
 }
+
 /**
  * Rotates the user current private key.
  * @param {string} passcode The users current passcode