+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The IronWeb SDK NPM releases follow standard [Semantic Versioning](https://semve
 **Note:** The patch versions of the IronWeb SDK will not be sequential and might jump by multiple numbers between sequential releases.
 
 ## v4.2.1
-- added `deleteDeviceByPublicSigningKeyWithJwt` to the User SDK. This allows someone with access to a JWT for a user to delete a device for that user without initializing the SDK (and creating a new browser device).
+- added `deleteDeviceByPublicSigningKeyWithJwt` to the SDK. This allows someone with access to a JWT for a user to delete a device for that user without initializing the SDK (and creating a new browser device).
 
 ## v4.1.1
 - added `listDevices()` to the User SDK. This lists all devices for the currently logged in user.